requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
tensorboard>=2.7.0
einops>=0.6.0
hydra-core>=1.1.0
omegaconf>=2.1.0
PyYAML>=5.4.0
h5py>=3.3.0
numba>=0.53.0
scipy>=1.7.0
jupyter>=1.0.0
notebook>=6.4.0
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.6b0
isort>=5.9.0
flake8>=3.9.0
wandb>=0.12.0
lightning>=2.0.0
transformers>=4.20.0  # For comparison with BERT baseline
causal-conv1d>=1.0.0  # For MAMBA implementation
mamba-ssm>=1.0.1  # MAMBA state space model
